pyproject.toml
[tool.poetry]
name = "greeting-card-bot"
version = "0.1.0"
description = ""
authors = ["Yurii Mironov <yuriymironov96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
